requirements.txt
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
filterpy==1.4.5
future==0.18.2
imageio==2.9.0
imutils==0.5.3
kiwisolver==1.3.1
loguru==0.5.3
matplotlib==3.3.3
networkx==2.5
numexpr==2.7.1
numpy==1.16.3
pandas==1.1.5
pandastable==0.12.2.post1
Pillow==8.0.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tbb==2020.3.254
tifffile
win32-setctime==1.0.3
xlrd==2.0.1
XlsxWriter==1.3.7
